Split checks into individual files
