+++ tests/checks/test_regulation.py
@@ -169,7 +169,7 @@ def test_validate_row_constraint_list_in_formula_issue_817():
 def test_validate_table_dimensions_num_rows():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(num_rows=42)],
+        checks=[checks.table_dimensions(num_rows=42)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"requiredNumRows": 42, "numberRows": 3}, "table-dimensions-error"]
@@ -189,7 +189,7 @@ def test_validate_table_dimensions_num_rows_declarative():
 def test_validate_table_dimensions_min_rows():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(min_rows=42)],
+        checks=[checks.table_dimensions(min_rows=42)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"minRows": 42, "numberRows": 3}, "table-dimensions-error"]
@@ -209,7 +209,7 @@ def test_validate_table_dimensions_min_rows_declarative():
 def test_validate_table_dimensions_max_rows():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(max_rows=2)],
+        checks=[checks.table_dimensions(max_rows=2)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"maxRows": 2, "numberRows": 3}, "table-dimensions-error"]
@@ -229,7 +229,7 @@ def test_validate_table_dimensions_max_rows_declarative():
 def test_validate_table_dimensions_num_fields():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(num_fields=42)],
+        checks=[checks.table_dimensions(num_fields=42)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"requiredNumFields": 42, "numberFields": 4}, "table-dimensions-error"]
@@ -249,7 +249,7 @@ def test_validate_table_dimensions_num_fields_declarative():
 def test_validate_table_dimensions_min_fields():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(min_fields=42)],
+        checks=[checks.table_dimensions(min_fields=42)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"minFields": 42, "numberFields": 4}, "table-dimensions-error"]
@@ -269,7 +269,7 @@ def test_validate_table_dimensions_min_fields_declarative():
 def test_validate_table_dimensions_max_fields():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(max_fields=2)],
+        checks=[checks.table_dimensions(max_fields=2)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"maxFields": 2, "numberFields": 4}, "table-dimensions-error"]
@@ -289,7 +289,7 @@ def test_validate_table_dimensions_max_fields_declarative():
 def test_validate_table_dimensions_no_limits():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions()],
+        checks=[checks.table_dimensions()],
     )
     assert report.flatten(["limits", "code"]) == []
 
@@ -305,7 +305,7 @@ def test_validate_table_dimensions_no_limits_declarative():
 def test_validate_table_dimensions_num_fields_num_rows_wrong():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(num_fields=3, num_rows=2)],
+        checks=[checks.table_dimensions(num_fields=3, num_rows=2)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"requiredNumFields": 3, "numberFields": 4}, "table-dimensions-error"],
@@ -327,7 +327,7 @@ def test_validate_table_dimensions_num_fields_num_rows_wrong_declarative():
 def test_validate_table_dimensions_num_fields_num_rows_correct():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(num_fields=4, num_rows=3)],
+        checks=[checks.table_dimensions(num_fields=4, num_rows=3)],
     )
     assert report.flatten(["limits", "code"]) == []
 
@@ -343,7 +343,7 @@ def test_validate_table_dimensions_num_fields_num_rows_correct_declarative():
 def test_validate_table_dimensions_min_fields_max_rows_wrong():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(min_fields=5, max_rows=2)],
+        checks=[checks.table_dimensions(min_fields=5, max_rows=2)],
     )
     assert report.flatten(["limits", "code"]) == [
         [{"minFields": 5, "numberFields": 4}, "table-dimensions-error"],
@@ -365,7 +365,7 @@ def test_validate_table_dimensions_min_fields_max_rows_wrong_declarative():
 def test_validate_table_dimensions_min_fields_max_rows_correct():
     report = validate(
         "data/table-limits.csv",
-        checks=[checks.regulation.table_dimensions(min_fields=4, max_rows=3)],
+        checks=[checks.table_dimensions(min_fields=4, max_rows=3)],
     )
     assert report.flatten(["limits", "code"]) == []
